Created the accounttransactiondetails.py, fundtransfer.py,transaction… (#3)
* Created the accounttransactiondetails.py, fundtransfer.py,transactiontype.py files

* Added the file mini_statement ,deleted the .pyc files and added in .gitignore, modified the files for debit and credit transactions

* removed the .pyc files

* Added the custom class exception in files , added .pyc in .gitignore,added mini statement

* Added the proper excpetion handling and response codes in the files

+++ app/view/branch_details.py
@@ -4,6 +4,7 @@ from app.model.bank_account import BranchDetails
 from flask import request
 from flask_restful import Resource
 from app.common.ResponseGenerator import ResponseGenerator
+from app.common.Exception import IdNotFound
 from flask_api import status
 from app.common.logging import *
 
@@ -31,8 +32,8 @@ class BranchData(Resource):
             return response.success_response()
         except Exception as error:
             logger.exception(error)
-            response = ResponseGenerator(data={}, message="Missing or sending incorrect data to create an activity",
-                                         success=False, status=status.HTTP_404_NOT_FOUND)
+            response = ResponseGenerator(data={}, message=error,
+                                         success=False, status=status.HTTP_400_BAD_REQUEST)
             return response.error_response()
 
 
@@ -53,8 +54,8 @@ class BranchData(Resource):
             return response.success_response()
         except Exception as error:
             logger.exception(error)
-            response = ResponseGenerator(data={}, message="Branch id not found", success=False,
-                                         status=status.HTTP_404_NOT_FOUND)
+            response = ResponseGenerator(data={}, message=error, success=False,
+                                         status=status.HTTP_400_BAD_REQUEST)
             return response.error_response()
 
 
@@ -74,15 +75,18 @@ class BranchInfo(Resource):
                                              status=status.HTTP_200_OK)
                 return response.success_response()
             else:
-                logger.exception("Branch id not found")
-                response = ResponseGenerator(data={}, message="Branch id not found", success=False,
-                                             status=status.HTTP_404_NOT_FOUND)
-                return response.error_response()
-        except Exception as error:
-            logger.exception(error)
-            response = ResponseGenerator(data={}, message="Branch id not found", success=False,
+                raise IdNotFound('id not found:{}'.format(id))
+        except IdNotFound as error:
+            logger.exception(error.message)
+            response = ResponseGenerator(data={}, message=error.message, success=False,
                                          status=status.HTTP_404_NOT_FOUND)
             return response.error_response()
+        except Exception as error:
+            logger.exception(error)
+            response = ResponseGenerator(data={}, message=error, success=False,
+                                         status=status.HTTP_400_BAD_REQUEST)
+            return response.error_response()
+
 
     def put(self, id):
 
@@ -104,9 +108,16 @@ class BranchInfo(Resource):
                 response = ResponseGenerator(data=output, message="Branch data updated successfully", success=True,
                                              status=status.HTTP_200_OK)
                 return response.success_response()
+            else:
+                raise IdNotFound('id not found:{}'.format(id))
+        except IdNotFound as error:
+            logger.exception(error.message)
+            response = ResponseGenerator(data={}, message=error.message, success=False,
+                                         status=status.HTTP_404_NOT_FOUND)
+            return response.error_response()
         except Exception as error:
             logger.exception(error)
-            response = ResponseGenerator(data={}, message="Missing or sending incorrect data to update an activity",
+            response = ResponseGenerator(data={}, message=error,
                                          success=False, status=status.HTTP_400_BAD_REQUEST)
             return response.error_response()
 
@@ -115,13 +126,21 @@ class BranchInfo(Resource):
         """delete the branch detail of selected branch  id"""
         try:
             branch = BranchDetails.query.get(id)
-            db.session.delete(branch)
-            db.session.commit()
-            logger.info("Branch details deleted successfully")
-            return "Branch details deleted successfully"
+            if branch:
+                db.session.delete(branch)
+                db.session.commit()
+                logger.info("Branch details deleted successfully")
+                return "Branch details deleted successfully"
+            else:
+                raise IdNotFound('id not found:{}'.format(id))
+        except IdNotFound as error:
+            logger.exception(error.message)
+            response = ResponseGenerator(data={}, message=error.message, success=False,
+                                         status=status.HTTP_404_NOT_FOUND)
+            return response.error_response()
         except Exception as error:
             logger.exception(error)
-            response = ResponseGenerator(data={}, message="branch id not found",
+            response = ResponseGenerator(data={}, message=error,
                                          success=False, status=status.HTTP_400_BAD_REQUEST)
             return response.error_response()
